requirements.txt
PySide6>=6.5.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
Pillow>=10.0.0
numpy>=1.24.0
pillow-heif>=0.13.0
faiss-cpu>=1.7.4
aiofiles>=23.0.0
watchdog>=3.0.0
networkx>=3.0